requirements.txt
asgiref==3.5.0
attrs==21.4.0
black==22.1.0
click==8.0.4
Django==4.0.3
djangorestframework==3.13.1
drf-spectacular==0.21.2
flake8==4.0.1
inflection==0.5.1
isort==5.10.1
jsonschema==4.4.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyrsistent==0.18.1
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
uritemplate==4.1.1
